requirements.txt
asgiref==3.3.1
attrs==20.3.0
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
colorzero==1.1
Django==3.1.3
django-bootstrap4==2.3.1
django-constance @ git+https://github.com/sflems/django-constance.git@6ac7e1b742f419f452bfa7940f3491b6f25d35b8
django-crispy-forms==1.10.0
django-jsonforms==1.1.2
django-picklefield==3.0.1
django-prettyjson==0.4.1
fastjsonschema==2.15.0
gpiozero==1.5.1
gunicorn==20.0.4
idna==2.10
importlib-metadata==2.1.1
jsonschema==3.2.0
psutil==5.8.0
pyrsistent==0.17.3
pytz==2020.5
redis==3.5.3
requests==2.25.1
RPi.GPIO==0.7.0
secret-key-generator==0.0.7
sh==1.14.1
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
standardjson==0.3.1
supervisor==4.2.2
urllib3==1.26.3
watchdog==2.0.2
whitenoise==5.2.0
zipp==3.4.0